layers: Fix crash on BeginCommandBuffer (INCOMPLETE)
This fixes a crash when non-null, garbage data is passed to
VkCommandBufferBeginInfo::pInheritanceInfo for a primary command buffer.
According to the spec:

pInheritanceInfo is a pointer to a VkCommandBufferInheritanceInfo
structure, used if commandBuffer is a secondary command buffer. If this
is a primary command buffer, then this value is ignored.

This fixes issue
[#2195](https://github.com/KhronosGroup/Vulkan-ValidationLayers/issues/2195).

TODO The following would need to be addressed in the generate scripts
before merging:
- ValidationObject::SantizeParameter:
  In layer_chassis_dispatch_generator.py would have to identify
  certain parameters to "sanitize"
- ValidationObject::StoreCommandBufferAllocationInfo:
  Would need to move AllocateCommandBuffers to the "manual section" of
  layer_chassis_generator.py

Change-Id: I6000d1e44065f1e028ab9735f64688ebe97b320a